feat(scanner): split SecurityScanner.scanners.ts + extend chmod evasion detection (SMI-5434, SMI-5433)
SMI-5434 — split SecurityScanner.scanners.ts (499→262 lines) ahead of the
500-line gate breach the chmod widening would have caused:
- SecurityScanner.compound.ts (NEW) — chmod+fetch compound detector (Section 2)
- SecurityScanner.pii.ts (NEW) — PII/entropy/placeholder scanning (Section 3)
- scanners.ts keeps Section 1 + re-exports the moved symbols (module surface
  unchanged; all 9 prior exports still importable). No behavior change.

SMI-5433 — widen OWNER_PERM_CHMOD + CHMOD_TARGET in compound.ts to catch three
evasion forms the old regex missed: comma-separated symbolic (a+w,o+x),
recursive flag (chmod -R 777), and assignment operator (u=rwx,g=rx). Core-only
— the edge scanner has no chmod detection (confirmed). H-6: route bare
CHMOD_TARGET.match / CHMOD_FETCH_CONTEXT.test through safeRegexTest (10k cap).

Adversarial-review fix (Opus): the planned regex used [ugoa]+ (one-or-more),
which silently DROPPED bare `chmod +x foo` — the most common make-executable
form in droppers — an FN regression. Restored [ugoa]* (zero-or-more); added
TP-4 regression-lock for bare `chmod +x`. Also corrected two TP/FP fixtures
that used world-writable forms (a+w / o+x) firing standalone-critical instead
of the intended compound-high path.

Verification (host): 91 core scanner tests pass incl. new chmod-evasion suite;
scanner-regression-guard passes with ZERO baseline delta (no corpus FPs);
pii-detection + exec tests resolve through the re-export layer; eslint +
prettier + audit:standards clean. Plans: docs/internal/implementation/
smi-5434-scanners-split.md, smi-5433-chmod-evasion-detection.md.
