+++ packages/core/src/security/scanner/SecurityScanner.scanners.ts
@@ -20,7 +20,6 @@ import {
   PROMPT_LEAKING_PATTERNS,
   DATA_EXFILTRATION_PATTERNS,
   PRIVILEGE_ESCALATION_PATTERNS,
-  PII_PATTERNS,
 } from './patterns.js'
 import { safeRegexTest, safeRegexCheck } from './regex-utils.js'
 import type { LineContext } from './SecurityScanner.helpers.js'
@@ -29,6 +28,11 @@ import {
   isDocumentationContext,
   isWithinInlineCode,
 } from './SecurityScanner.helpers.js'
+import {
+  looksLikePlaceholderSecret,
+  shannonEntropy,
+  scanPiiPatterns,
+} from './SecurityScanner.pii.js'
 
 /**
  * SMI-5359 Wave 4 (MF-2): a `.env` reference is an active read/exfiltration only when
@@ -256,244 +260,7 @@ export function scanPrivilegeEscalation(
   return findings
 }
 
-/**
- * SMI-5424 PR2: owner-permission chmod is a COMPOUND signal, not standalone —
- * `chmod 755 ./bin/cli` / `chmod +x build.sh` previously false-fired
- * privilege_escalation:critical. It now emits HIGH only when a fetch COMMAND
- * (curl/wget/git-clone/npx-to-URL) is within ±1 line OR the chmod target is the
- * DOWNLOAD DESTINATION of a fetch command anywhere (distance-independent, so filler
- * lines can't evade the ±1 window) — the "download a payload, chmod it, run it" shape.
- * Kills the FP while PRESERVING the chmod co-signal escalateCodeExecution needs (it
- * only accepts high/critical non-doc co-signals). World-writable / setuid chmod stay
- * standalone-critical in PRIVILEGE_ESCALATION_PATTERNS; `alreadyFlaggedLines` prevents
- * double-emit. SMI-5431: "destination" covers explicit (-o/-O/--output<space>/>) AND
- * implicit (wget no -O / git clone / curl --output=) targets; a bare `curl <url>` GET
- * writes to STDOUT so it is NOT correlated (the URL-path-segment FP a prior review caught).
- * The ONLY uncaught residual: a SPACED `curl … | bash` (no filename) + a NON-adjacent chmod.
- */
-const OWNER_PERM_CHMOD = /\bchmod\s+(?:[0-7]{3,4}|[ugoa]*\+x)\b/i
-// FIX-1: actual fetch COMMANDS only — bare prose tokens (`# downloaded`, `See https://…`,
-// `npx tool init`) false-fired next to a chmod. curl/wget/git-clone, and npx only with a URL.
-const CHMOD_FETCH_CONTEXT = /\b(?:curl|wget)\b|\bgit\s+clone\b|\bnpx\b[^\n]{0,80}https?:\/\//i
-// FIX-2: capture the chmod target path so a fetch command anywhere correlates by basename.
-const CHMOD_TARGET = /\bchmod\s+(?:[0-7]{3,4}|[ugoa]*\+x)\s+(\S+)/i
-function escapeRegExp(s: string): string {
-  return s.replace(/[.*+?^${}()|[\]\\]/g, '\\$&')
-}
-// SMI-5431: the IMPLICIT download destination of a fetch command — the file written with
-// NO explicit -o/-O/--output<space>/> redirect: `wget <url>` (no -O/-o) → URL last segment;
-// `git clone <url>` → repo dir (minus `.git`); `curl --output=<file>` (equals form, missed by
-// the explicit regex). A bare `curl <url>` GET writes to STDOUT → '' (never correlates). ReDoS-safe.
-function implicitDownloadBasename(line: string): string {
-  const lastSegment = (urlAfterScheme: string): string => {
-    const noFrag = urlAfterScheme.split(/[?#]/)[0]
-    const slash = noFrag.indexOf('/') // first slash = end of host
-    if (slash < 0) return '' // host only -> wget writes index.html
-    const path = noFrag.slice(slash + 1).replace(/\/+$/, '')
-    return path === '' ? '' : (path.split('/').pop() ?? '')
-  }
-  const wget = line.match(/\bwget\b(?![^\n]{0,200}\s-[oO]\b)[^\n]{0,200}?https?:\/\/(\S{1,400})/i)
-  if (wget) return lastSegment(wget[1])
-  const clone = line.match(/\bgit\s+clone\b[^\n]{0,200}?https?:\/\/(\S{1,400})/i)
-  if (clone) return lastSegment(clone[1]).replace(/\.git$/i, '')
-  const curlEq = line.match(/\bcurl\b[^\n]{0,200}?--output=['"]?(\S{1,400})/i)
-  if (curlEq) return curlEq[1].replace(/['"]/g, '').split('/').pop() ?? ''
-  return ''
-}
-
-export function scanChmodFetchCompound(
-  content: string,
-  alreadyFlaggedLines: ReadonlySet<number>,
-  lineContexts?: LineContext[]
-): SecurityFinding[] {
-  const findings: SecurityFinding[] = []
-  const lines = content.split('\n')
-  const contexts = lineContexts ?? analyzeMarkdownContext(content)
-  // FIX-2: lines carrying a fetch command, for distance-independent correlation.
-  const fetchLines = lines.filter((l) => CHMOD_FETCH_CONTEXT.test(l))
-
-  lines.forEach((line, index) => {
-    const lineNumber = index + 1
-    // World-writable / setuid already emitted critical for this line — skip.
-    if (alreadyFlaggedLines.has(lineNumber)) return
-    const match = safeRegexTest(OWNER_PERM_CHMOD, line)
-    if (!match) return
-    // Bounded ±1-line window for the download-then-chmod adjacency.
-    const window = [lines[index - 1] ?? '', line, lines[index + 1] ?? ''].join('\n')
-    const adjacentFetch = CHMOD_FETCH_CONTEXT.test(window)
-    // FIX-2 + SMI-5431: correlate the chmod target basename (≥3 chars) against a fetch
-    // command's DOWNLOAD DESTINATION anywhere — explicit (-o/-O/--output<space>/>, with an
-    // optional leading path) via regex, OR implicit (wget/git-clone/curl --output=) via
-    // exact-token equality. Anchored on the destination, NOT basename-anywhere, so a URL
-    // path / query / header value (governance FP class) and a bare curl GET do not correlate.
-    let correlated = false
-    const tm = line.match(CHMOD_TARGET)
-    if (tm) {
-      const base = tm[1].replace(/['"]/g, '').split('/').pop() ?? ''
-      if (base.length >= 3) {
-        const re = new RegExp(
-          `(?:-o|-O|--output|>>?)\\s*['"]?(?:[^\\s'"]*/)?${escapeRegExp(base)}(?:[\\s'"?]|$)`
-        )
-        correlated = fetchLines.some((l) => re.test(l) || implicitDownloadBasename(l) === base)
-      }
-    }
-    if (!adjacentFetch && !correlated) return // benign standalone chmod — no finding
-
-    const ctx = contexts[index]
-    const inInlineCode = ctx?.isInlineCode && isWithinInlineCode(line, match.index ?? 0)
-    const inDocContext = ctx ? isDocumentationContext(ctx) || inInlineCode : false
-    findings.push({
-      type: 'privilege_escalation',
-      // HIGH (not critical): enough to trip Gate-A AND serve as an
-      // escalateCodeExecution co-signal, without re-introducing a critical FP.
-      severity: inDocContext ? 'low' : 'high',
-      message: `chmod of a fetched/downloaded file (compound with a download verb): "${match[0]}"`,
-      location: line.trim().slice(0, 100),
-      lineNumber,
-      category: 'privilege_escalation',
-      inDocumentationContext: inDocContext,
-      confidence: inDocContext ? 'low' : 'high',
-    })
-  })
-
-  return findings
-}
-
-/**
- * SMI-5420: credential PII pattern indices (api/secret/access key, provider
- * tokens, password) whose matched VALUE can be a documentation placeholder.
- * Excludes email (7), SSN (8), and the private-key marker (9) — those have no
- * placeholder-secret failure mode.
- */
-const CREDENTIAL_PII_INDICES = new Set([0, 1, 2, 3, 4, 5, 6, 10])
-
-/**
- * SMI-5420: named-placeholder markers that indicate an example, not a real
- * secret. Intentionally NO `X{4,}` rule — it would test the raw match string and
- * short-circuit before the entropy check, downgrading a REAL high-entropy secret
- * that coincidentally contains `xxxx` (governance FN). An all-repeated-char value
- * (e.g. `xxxx…`) is caught by the `/^(.)\1+$/` check in looksLikePlaceholderSecret
- * instead, and partial-repeat low-variety values by the entropy floor.
- *
- * SMI-5423: the SHORT markers (FAKE/DUMMY/SAMPLE/YOUR, ≤6 chars) are guarded with
- * a negative lookbehind `(?<![A-Za-z0-9])` so they only match as a delimited token
- * (`FAKE_KEY`, `<FAKE>`, value-start) — NOT mid-random-string (`k7FAKE1abc`), which
- * is the same raw-match short-circuit FN class as the removed `X{4,}`. Longer
- * markers (EXAMPLE/PLACEHOLDER/CHANGEME/REDACTED, 7+ chars) stay unbounded: their
- * coincidence probability is negligible AND `AKIA…7EXAMPLE` needs EXAMPLE to match
- * mid-token.
- *
- * Accepted tradeoff (SMI-5423 governance): a digit-immediately-prefixed token like
- * `1FAKE_KEY` fails the lookbehind and scores critical rather than low. This is the
- * FP-SAFE direction (over-flag a rare contrived placeholder) — strictly preferable
- * in a security scanner to the FN it replaces (a real secret downgraded); and a
- * longer such value falls under the entropy floor anyway.
- */
-const PLACEHOLDER_SECRET_RE =
-  /EXAMPLE|(?<![A-Za-z0-9])YOUR[_-]?|PLACEHOLDER|CHANGE[_-]?ME|(?<![A-Za-z0-9])DUMMY|(?<![A-Za-z0-9])FAKE|(?<![A-Za-z0-9])SAMPLE|REDACTED|INSERT[_-]|\.\.\.|<[^>]+>/i
-
-/**
- * SMI-5420: minimum Shannon entropy (bits/char) for a value to read as a real secret.
- * SMI-5424 PR2 (accepted tradeoff): a hardcoded credential value shorter than ~8 chars
- * cannot reach 3.0 bits/char (its max entropy is log2(len) < 3.0), so it falls below
- * this floor and is treated as a placeholder by design — a sub-8-char literal is not a
- * credible credential, and the 20+char PII_PATTERNS rule is the real-credential detector.
- */
-const SECRET_ENTROPY_FLOOR = 3.0
-
-/** SMI-5420: Shannon entropy (bits per character) of a string. */
-export function shannonEntropy(s: string): number {
-  if (!s) return 0
-  const freq = new Map<string, number>()
-  for (const ch of s) freq.set(ch, (freq.get(ch) ?? 0) + 1)
-  let h = 0
-  for (const c of freq.values()) {
-    const p = c / s.length
-    h -= p * Math.log2(p)
-  }
-  return h
-}
-
-/**
- * SMI-5420: extract the secret token from a credential match by stripping a
- * leading `<key>:`/`<key>=` assignment prefix and surrounding quotes.
- */
-function extractSecretValue(match: string): string {
-  return match
-    .replace(/^[^:=]*[:=]\s*/, '')
-    .replace(/^['"]|['"]$/g, '')
-    .trim()
-}
-
-/**
- * SMI-5420: a credential match is a documentation placeholder (not a real leaked
- * secret) when it carries a named placeholder marker, is a single repeated
- * character, or its value has sub-secret Shannon entropy. Such matches must NOT
- * emit critical/high severity — the batch trust-scorer (trust-scorer.ts) and the
- * install gate quarantine on severity, so an example secret would falsely flag.
- */
-export function looksLikePlaceholderSecret(match: string): boolean {
-  if (PLACEHOLDER_SECRET_RE.test(match)) return true
-  const value = extractSecretValue(match)
-  if (value.length === 0) return false
-  if (/^(.)\1+$/.test(value)) return true
-  return shannonEntropy(value) < SECRET_ENTROPY_FLOOR
-}
-
-/** SMI-3864: Detect PII patterns. Email in YAML frontmatter gets low severity. */
-export function scanPiiPatterns(content: string, lineContexts?: LineContext[]): SecurityFinding[] {
-  const findings: SecurityFinding[] = []
-  const lines = content.split('\n')
-  const contexts = lineContexts ?? analyzeMarkdownContext(content)
-  let frontmatterEnd = -1
-  if (lines[0]?.trim() === '---') {
-    for (let i = 1; i < lines.length; i++) {
-      if (lines[i].trim() === '---') {
-        frontmatterEnd = i
-        break
-      }
-    }
-  }
-  const emailPatternIndex = 7
-  lines.forEach((line, index) => {
-    const ctx = contexts[index]
-    const inFrontmatter = index > 0 && index < frontmatterEnd
-    for (let pi = 0; pi < PII_PATTERNS.length; pi++) {
-      const pattern = PII_PATTERNS[pi]
-      const match = safeRegexTest(pattern, line)
-      if (match) {
-        const inInlineCode = ctx?.isInlineCode && isWithinInlineCode(line, match.index ?? 0)
-        const inDocContext = ctx ? isDocumentationContext(ctx) || inInlineCode : false
-        const isEmailPattern = pi === emailPatternIndex
-        const isAuthorLine = /^\s*(?:author|contact|support|email)\s*:/i.test(line)
-        const inEmailSafeContext = isEmailPattern && (inFrontmatter || isAuthorLine)
-        let severity: 'low' | 'medium' | 'high' | 'critical'
-        if (inEmailSafeContext) severity = 'low'
-        else if (inDocContext) severity = 'medium'
-        else if (pi <= 2 || pi === 9) severity = 'critical'
-        else severity = 'high'
-        let confidence: FindingConfidence = inDocContext || inEmailSafeContext ? 'low' : 'high'
-        // SMI-5420: a credential match that reads as a documentation placeholder
-        // (named placeholder, repeated char, or low entropy) must not emit
-        // critical/high — the batch trust-scorer quarantines on severity, so an
-        // example secret like `api_key: "YOUR_API_KEY_HERE"` would falsely flag.
-        if (CREDENTIAL_PII_INDICES.has(pi) && looksLikePlaceholderSecret(match[0])) {
-          severity = 'low'
-          confidence = 'low'
-        }
-        findings.push({
-          type: 'pii',
-          severity,
-          message: `PII detected: ${match[0].slice(0, 40)}${match[0].length > 40 ? '...' : ''}`,
-          location: line.trim().slice(0, 100),
-          lineNumber: index + 1,
-          category: 'pii',
-          inDocumentationContext: inDocContext || inEmailSafeContext,
-          confidence,
-        })
-        break
-      }
-    }
-  })
-  return findings
-}
+/** Implementation in SecurityScanner.compound.ts (SMI-5434 split). */
+export { scanChmodFetchCompound } from './SecurityScanner.compound.js'
+/** Implementation in SecurityScanner.pii.ts (SMI-5434 split). */
+export { shannonEntropy, looksLikePlaceholderSecret, scanPiiPatterns }
